refactor: restructure examples (#221)
* refactor: restructure examples

* fix: put back wabt as dev-dependencies

* fix: clippy warnings on unused fields
