+++ examples/cli.yml
