+++ examples/example_contract.rs.DEPRECATED
