+++ examples/example_print_metadata.rs
@@ -21,14 +21,11 @@ extern crate clap;
 
 use clap::App;
 
-use node_template_runtime::{Block, Header};
 use sp_core::sr25519;
-use sp_runtime::generic::SignedBlock as SignedBlockG;
-use std::sync::mpsc::channel;
-use substrate_api_client::rpc::WsRpcClient;
-use substrate_api_client::Api;
 
-type SignedBlock = SignedBlockG<Block>;
+use std::convert::TryFrom;
+use substrate_api_client::rpc::WsRpcClient;
+use substrate_api_client::{Api, Metadata};
 
 fn main() {
     env_logger::init();
@@ -37,46 +34,25 @@ fn main() {
     let client = WsRpcClient::new(&url);
     let api = Api::<sr25519::Pair, _>::new(client).unwrap();
 
-    let head = api.get_finalized_head().unwrap().unwrap();
+    let meta = Metadata::try_from(api.get_metadata().unwrap()).unwrap();
 
+    meta.print_overview();
+    meta.print_pallets();
+    meta.print_pallets_with_calls();
+    meta.print_pallets_with_events();
+    meta.print_pallets_with_errors();
+    meta.print_pallets_with_constants();
+
+    // print full substrate metadata json formatted
     println!(
-        "Genesis block: \n {:?} \n",
-        api.get_block_by_num::<Block>(Some(0)).unwrap()
-    );
-
-    println!("Finalized Head:\n {} \n", head);
-
-    let h: Header = api.get_header(Some(head)).unwrap().unwrap();
-    println!("Finalized header:\n {:?} \n", h);
-
-    let b: SignedBlock = api.get_signed_block(Some(head)).unwrap().unwrap();
-    println!("Finalized signed block:\n {:?} \n", b);
-
-    println!(
-        "Latest Header: \n {:?} \n",
-        api.get_header::<Header>(None).unwrap()
-    );
-
-    println!(
-        "Latest block: \n {:?} \n",
-        api.get_block::<Block>(None).unwrap()
-    );
-
-    println!("Subscribing to finalized heads");
-    let (sender, receiver) = channel();
-    api.subscribe_finalized_heads(sender).unwrap();
-
-    for _ in 0..5 {
-        let head: Header = receiver
-            .recv()
-            .map(|header| serde_json::from_str(&header).unwrap())
-            .unwrap();
-        println!("Got new Block {:?}", head);
-    }
+        "{}",
+        Metadata::pretty_format(&api.get_metadata().unwrap())
+            .unwrap_or_else(|| "pretty format failed".to_string())
+    )
 }
 
 pub fn get_node_url_from_cli() -> String {
-    let yml = load_yaml!("../../src/examples/cli.yml");
+    let yml = load_yaml!("cli.yml");
     let matches = App::from_yaml(yml).get_matches();
 
     let node_ip = matches.value_of("node-server").unwrap_or("ws://127.0.0.1");
